--- ingestion_tools/scripts/data_validation/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-# Ignore tests in a data_validation folder (asterisk before to glob if in scripts/ or ingestion_tools/ folder)
-addopts = "--ignore=results"
