+++ ingestion_tools/scripts/data_validation/standardized/pytest.ini
@@ -1,0 +1,16 @@
+[pytest]
+# Ignore tests in a data_validation folder (asterisk before to glob if in scripts/ or ingestion_tools/ folder)
+addopts = "--ignore=results"
+
+markers =
+    alignment: Tests concerning the alignment data.
+    annotation: Tests concerning the annotation data.
+    dataset: Tests concerning the dataset.
+    deposition: Tests concerning the deposition data.
+    frame: Tests concerning the frames.
+    gain: Tests concerning the gain files.
+    run: Tests concerning the runs.
+    tiltseries: Tests concerning the tiltseries.
+    tilt_angles: Tests concerning the tilt angle (spans multiple entities: .tlt, .rawtlt, .mdoc, tiltseries_metadata.json, frames).
+    tomogram: Tests concerning the tomogram.
+    voxel_spacing: Tests concerning voxel spacings.
