+++ ingestion_tools/scripts/data_validation/standardized/tests/annotation/helper_point.py
