+++ ingestion_tools/scripts/data_validation/standardized/tests/annotation/test_instancesegmentation_annotation.py
@@ -2,7 +2,8 @@ from typing import Dict, List
 
 import allure
 import pytest
-from data_validation.tests.annotation.helper_point import contained_in_tomo
+
+from data_validation.standardized.tests.annotation.helper_point import contained_in_tomo
 
 
 @pytest.mark.annotation
