+++ ingestion_tools/scripts/data_validation/standardized/tests/annotation/test_metadata_annotation.py
@@ -2,10 +2,10 @@ from typing import Dict
 
 import allure
 import pytest
-from data_validation.tests.helper_metadata import basic_metadata_check
-from data_validation.tests.test_deposition import HelperTestDeposition
 
 from common.fs import FileSystemApi
+from data_validation.standardized.tests.helper_metadata import basic_metadata_check
+from data_validation.standardized.tests.test_deposition import HelperTestDeposition
 
 
 @pytest.mark.annotation
