+++ ingestion_tools/scripts/data_validation/standardized/tests/annotation/test_orientedpoint_annotation.py
@@ -3,7 +3,8 @@ from typing import Dict, List
 import allure
 import numpy as np
 import pytest
-from data_validation.tests.annotation.helper_point import contained_in_tomo, point_count_consistent
+
+from data_validation.standardized.tests.annotation.helper_point import contained_in_tomo, point_count_consistent
 
 
 @pytest.mark.annotation
