+++ ingestion_tools/scripts/data_validation/standardized/tests/helper_images.py
