+++ ingestion_tools/scripts/data_validation/standardized/tests/test_deposition.py
@@ -6,10 +6,9 @@ by other test classes to validate their corresponding deposition's metadata.
 import json
 from typing import Dict, Union
 
-from data_validation.tests.helper_images import check_photo_valid
-from data_validation.tests.helper_metadata import basic_metadata_check
-
 from common.fs import FileSystemApi
+from data_validation.standardized.tests.helper_images import check_photo_valid
+from data_validation.standardized.tests.helper_metadata import basic_metadata_check
 
 
 class HelperTestDeposition:
